feat: model alunos
